+++ forge.config.ts
@@ -1,0 +1,36 @@
+import { MakerDeb } from '@electron-forge/maker-deb';
+import { MakerRpm } from '@electron-forge/maker-rpm';
+import { MakerSquirrel } from '@electron-forge/maker-squirrel';
+import { MakerZIP } from '@electron-forge/maker-zip';
+import { WebpackPlugin } from '@electron-forge/plugin-webpack';
+
+import { mainConfig } from './webpack.main.config';
+import { rendererConfig } from './webpack.renderer.config';
+
+import type { ForgeConfig } from "@electron-forge/shared-types";
+
+const config: ForgeConfig = {
+  packagerConfig: {},
+  rebuildConfig: {},
+  makers: [new MakerSquirrel({}), new MakerZIP({}, ["darwin"]), new MakerRpm({}), new MakerDeb({})],
+  plugins: [
+    new WebpackPlugin({
+      mainConfig,
+      renderer: {
+        config: rendererConfig,
+        entryPoints: [
+          {
+            html: "./src/renderer/index.html",
+            js: "./src/renderer/index.ts",
+            name: "main_window",
+            preload: {
+              js: "./src/preload/index.ts",
+            },
+          },
+        ],
+      },
+    }),
+  ],
+};
+
+export default config;